sources/SDN_CLOUD_1-master/dashboard/requirements-dashboard.txt
Flask>=3.0,<4.0
requests>=2.31,<3.0
openstacksdk>=1.5,<2.0
PyYAML>=6.0,<7.0
